Move :generator:iconpack to sdk
