--- components/generator/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/generator/iconpack/IconPackGeneratorConfig.kt
@@ -1,11 +1,0 @@
-package io.github.composegears.valkyrie.generator.iconpack
-
-import io.github.composegears.valkyrie.generator.core.IconPack
-
-data class IconPackGeneratorConfig(
-    val packageName: String,
-    val iconPack: IconPack,
-    val useExplicitMode: Boolean,
-    val indentSize: Int,
-    val license: String? = null,
-)
