--- components/generator/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/generator/iconpack/IconPackSpecOutput.kt
@@ -1,6 +1,0 @@
-package io.github.composegears.valkyrie.generator.iconpack
-
-data class IconPackSpecOutput(
-    val content: String,
-    val name: String,
-)
