--- components/generator/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/generator/iconpack/LicenseFormatter.kt
@@ -1,6 +1,0 @@
-package io.github.composegears.valkyrie.generator.iconpack
-
-internal fun String.asBlockComment(): String {
-    if (startsWith("/*") && endsWith("*/")) return this
-    return "/*\n${prependIndent(" * ")}\n */"
-}
