+++ sdk/generator/kt/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/iconpack/IconPackGenerator.kt
@@ -1,0 +1,5 @@
+package io.github.composegears.valkyrie.sdk.generator.kt.iconpack
+
+expect object IconPackGenerator {
+    fun create(config: IconPackGeneratorConfig): IconPackSpecOutput
+}
