+++ sdk/generator/kt/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/iconpack/IconPackSpecOutput.kt
@@ -1,0 +1,6 @@
+package io.github.composegears.valkyrie.sdk.generator.kt.iconpack
+
+data class IconPackSpecOutput(
+    val content: String,
+    val name: String,
+)
