+++ sdk/generator/kt/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/iconpack/LicenseFormatter.kt
@@ -1,0 +1,6 @@
+package io.github.composegears.valkyrie.sdk.generator.kt.iconpack
+
+internal fun String.asBlockComment(): String {
+    if (startsWith("/*") && endsWith("*/")) return this
+    return "/*\n${prependIndent(" * ")}\n */"
+}
