+++ sdk/generator/kt/iconpack/src/jvmMain/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/iconpack/IconPackGenerator.jvm.kt
@@ -1,0 +1,9 @@
+package io.github.composegears.valkyrie.sdk.generator.kt.iconpack
+
+import io.github.composegears.valkyrie.sdk.generator.kt.iconpack.internal.IconPackFileSpec
+
+actual object IconPackGenerator {
+    actual fun create(config: IconPackGeneratorConfig): IconPackSpecOutput {
+        return IconPackFileSpec(config).createSpec()
+    }
+}
