+++ sdk/generator/kt/iconpack/src/wasmJsMain/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/iconpack/IconPackGenerator.wasmJs.kt
@@ -1,0 +1,50 @@
+package io.github.composegears.valkyrie.sdk.generator.kt.iconpack
+
+import io.github.composegears.valkyrie.generator.core.IconPack
+
+actual object IconPackGenerator {
+
+    actual fun create(config: IconPackGeneratorConfig): IconPackSpecOutput {
+        val content = buildString {
+            if (config.license != null) {
+                appendLine(config.license.asBlockComment())
+                appendLine()
+            }
+            appendLine("package ${config.packageName}")
+            appendLine()
+
+            val iconPack = createIconPack(
+                iconPack = config.iconPack,
+                indentSize = config.indentSize,
+                isExplicit = config.useExplicitMode,
+            ).clearBlankLines()
+            appendLine(iconPack)
+        }
+
+        return IconPackSpecOutput(name = config.iconPack.data, content = content)
+    }
+
+    private fun createIconPack(iconPack: IconPack, indentSize: Int, isExplicit: Boolean): String {
+        return buildString {
+            if (isExplicit) append("public ")
+            append("object ${iconPack.data}")
+
+            if (iconPack.children.isNotEmpty()) {
+                appendLine(" {")
+                val lastIndex = iconPack.children.lastIndex
+                iconPack.children.forEachIndexed { i, nestedPack ->
+                    val pack = createIconPack(
+                        iconPack = nestedPack,
+                        indentSize = indentSize,
+                        isExplicit = isExplicit,
+                    )
+                    appendLine(pack.prependIndent(indent = " ".repeat(indentSize)))
+                    if (i != lastIndex) appendLine()
+                }
+                append("}")
+            }
+        }
+    }
+
+    private fun String.clearBlankLines(): String = this.lines().joinToString("\n") { it.takeUnless { it.isBlank() }.orEmpty() }
+}
