+++ settings.gradle.kts
@@ -72,7 +72,6 @@ include("tools:idea-plugin")
 
 include("components:generator:core")
 include("components:generator:kmp:imagevector")
-include("components:generator:iconpack")
 include("components:generator:jvm:imagevector")
 include("components:parser:jvm:svg")
 include("components:parser:jvm:xml")
@@ -90,6 +89,7 @@ include("sdk:core:tree")
 include("sdk:core:xml")
 include("sdk:generator:kt:common:util")
 include("sdk:generator:kt:common:poet-extensions")
+include("sdk:generator:kt:iconpack")
 include("sdk:generator:xml")
 include("sdk:intellij:psi:iconpack")
 include("sdk:intellij:psi:imagevector")
